Rough refactoring complete.
