+++ src/modules/results.ts
@@ -1,26 +1,26 @@
-import { FrameInt, PlayerInt } from "../types/types";
+/*
+  This module is separate from all other .ts modules. It:
+  1. Reads prepared HTML string from Session Storage.
+  2. Displays it.
+*/
 
 // DOM Elements
 const tableBody = document.querySelector('.table__body');
 
-console.log('This should work on table page only.');
-
-const getDataFromSessionStorage = () => {
-  const data = window.sessionStorage.getItem( 'playerResults' );
-  let playerData: string = `No results found.`;
-
-  if ( data !== null ) {
-    playerData = JSON.parse( data );
-  }
-
-  return playerData;
-}
-
-// displayResults( playerData );
-
-(function display () {
+export const display = (() => {
   const tableRows = getDataFromSessionStorage();
   if ( tableBody !== null ) {
     tableBody.innerHTML = tableRows;
   };
 })();
+
+function getDataFromSessionStorage () {
+  const data = window.sessionStorage.getItem( 'playerResults' );
+  let playerData: string = `No results found.`;
+
+  if ( data !== null ) {
+    playerData = JSON.parse( data );
+  };
+
+  return playerData;
+};
